Fixed high memory usage, added UNHEX macro
The high memory usage was caused by not freeing the surface and texture
of the status bar text; now in every main loop iteration the a surface
and texture is created to draw the statusbar then its freen and
reallocated in the next iteration.

Minor updates on the code has been done as well.