Fixed high CPU usage
Fixed high CPU usage by creating two functions to delay and pause the
sorting algorithm being run, sort_delay() and sort_pause() respectively;
both of this functions use a small delay in between the loop used to
wait until a condition breaks it.

The drw_element_color() function had a small change too, the line drawn
around the rect was moved by one pixel to the top and it was in the
wrong, side; now is drawn correctly.